+++ dbt/include/spark/macros/materializations/incremental/validate.sql
@@ -28,13 +28,13 @@
     Invalid incremental strategy provided: {{ raw_strategy }}
     You can only choose this strategy when file_format is set to 'delta' or 'hudi'
   {%- endset %}
-  
+
   {% set invalid_insert_overwrite_delta_msg -%}
     Invalid incremental strategy provided: {{ raw_strategy }}
     You cannot use this strategy when file_format is set to 'delta'
     Use the 'append' or 'merge' strategy instead
   {%- endset %}
-  
+
   {% set invalid_insert_overwrite_endpoint_msg -%}
     Invalid incremental strategy provided: {{ raw_strategy }}
     You cannot use this strategy when connecting via endpoint